"Add hospital routes for adding and getting hospital details"
This commit adds two new routes for handling hospital-related functionality. The first route, `addHospital`, is a POST request to add a new hospital to the system. The second route, `getHospital/:id`, is a GET request to retrieve the details of a hospital with a given ID. These routes use the `addHospital` and `getHospitalDetails` functions from the `hospitalController`.